requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==1.11.4
django-allauth==0.33.0
django-appconf==1.0.2
django-braces==1.12.0
django-crispy-forms==1.6.1
django-imagekit==4.0.1
django-js-asset==0.1.1
django-model-utils==3.0.0
django-mptt==0.8.7
django-notifications-hq==1.3
django-redis==4.8.0
django-suit==0.2.25
django-taggit==0.22.1
gevent==1.2.2
greenlet==0.4.12
idna==2.6
jsonfield==2.0.2
oauthlib==2.0.2
olefile==0.44
pigar==0.7.1
pilkit==2.0
Pillow==4.2.1
psycopg2==2.7.3.1
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
websockets==3.2
